[PATCH] Allow LLDD's to fail slave alloc (non-existent slave)
In the cases where an LLDD knows that a lun does not exist, if it
fails slave_alloc(), the midlayer reports an error message indicating
slave_alloc failed.

This patch allows the LLDD to reject the slave_alloc() call if it
knows the lun doesn't exist, and skips the reporting of the
slave_alloc failure message. This avoids jumping through hoops in the
LLDD to avoid error messages from slave_alloc. Also saves cycles in
that the probe is also cancelled for the non-existent device.

Signed-off-by: James Bottomley <James.Bottomley@SteelEye.com>