[PATCH] ppc64: Remote some userland-only stuff from kernel header
include/asm-ppc64/systemcfg.h defines a structure that currently gets
exposed to userland via /proc, and which I intend to deprecate in the
near future once I have better alternatives available.

In the meantime, this patch removes a bunch of stuff from this file that
were only defined for non-__KERNEL__, like an inline function for
getting to that struture via /proc, and some CPU & platform type
definitions that were duplicates of the ones in asm-ppc64/processor.h.
These things have nothing to do in a kernel header.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>