[PATCH] Fix booting on some recent G5s
Some recent G5s have a problem with PCI/HT probing.  They crash (machine
check) during the probe of some slot numbers, it seems to be related to
some functions beeing disabled by the firmware inside the K2 ASIC.

This patch limits the config space accesses to devices that are present
in the OF device-tree.  This fixes the problem and shouldn't "add" any
limitation.  If you plug a "random" PCI card with no OF driver, the
firmware will still build a node for it with the default set of
properties created from the config space. 

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>