[PATCH] smc91x: simplify register bank usage
From: Nicolas Pitre <nico@cam.org>

Let's avoid all that crap of preserving the bank to restore it all the time if
that's not necessary.  The rule is simple: bank 2 is the only allowed bank
when the chip is active and unlocked.  Anything that needs to change the bank
to something other than bank 2 must lock against any concurrent access and
return to bank 2 before unlocking.

Signed-off-by: Nicolas Pitre <nico@cam.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>